+++ frontend/src/js/components/SpecialPromotion.jsx
@@ -1,26 +1,11 @@
 import React, {useState} from 'react';
 import {specialPromotionIcon} from '../constants/iconPath.js';
 import {BestSideDish} from './BestSideDish.jsx';
-import {goodsData} from '../data/goods.js';
+import {TabList} from './TabList.jsx';
 import '../../css/SpecialPromotion.css';
 
 function SpecialPromotion() {
   const [tabState, setTabState] = useState('풍부한 고기 반찬');
-
-  const handleTabState = ({target}) => {
-    setTabState(target.title);
-  };
-
-  const tabList = goodsData.map(element => (
-    <li
-      className={tabState === element.tab.title ? 'selected' : ''}
-      key={element.id}
-      title={element.tab.title}
-      onClick={handleTabState}
-    >
-      {element.tab.title}
-    </li>
-  ));
 
   return (
     <div className="specialPromotion">
@@ -30,7 +15,9 @@ function SpecialPromotion() {
           <p>한 번 주문하면 두 번 반하는 반찬</p>
         </h2>
         <h4 className="tab">
-          <ul className="tabList">{tabList}</ul>
+          <ul className="tabList">
+            <TabList tabState={tabState} setTabState={setTabState} />
+          </ul>
         </h4>
         <div className="bestSideDishContainer">
           <ul className="bestSideDishList">
